+++ stream-compositions/services/product-composition-service/src/main/java/com/backbase/stream/compositions/product/core/service/impl/TransactionIngestionServiceImpl.java
@@ -11,9 +11,11 @@ import com.backbase.stream.compositions.product.core.service.TransactionIngestio
 import com.backbase.stream.compositions.transaction.client.TransactionCompositionApi;
 import com.backbase.stream.compositions.transaction.client.model.TransactionPullIngestionRequest;
 import java.util.List;
+import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 import reactor.core.scheduler.Schedulers;
 
@@ -31,7 +33,9 @@ public class TransactionIngestionServiceImpl implements TransactionIngestionServ
     @Override
     public Mono<?> ingestTransactions(TransactionPullIngestionRequest res) {
         if (properties.getChains().getTransactionManager().getEnabled()) {
-            return refreshTransactions(res);
+            return Mono.just(res)
+                .map(r -> new ArrangementItem().arrangementId(r.getExternalArrangementId()))
+                .flatMap(r -> refreshTransactions(List.of(r)));
         }
         return transactionCompositionApi.pullTransactions(res)
             .doOnNext(response -> log.debug("Response from Transaction Composition: {}",
@@ -39,11 +43,27 @@ public class TransactionIngestionServiceImpl implements TransactionIngestionServ
     }
 
     @Override
+    public Flux<?> ingestTransactions(Flux<TransactionPullIngestionRequest> res) {
+        if (properties.getChains().getTransactionManager().getEnabled()) {
+            if (properties.getChains().getTransactionManager().getSplitPerArrangement()) {
+                return res.flatMap(this::ingestTransactions,
+                    properties.getChains().getTransactionManager().getConcurrency());
+            }
+            return res.map(r -> new ArrangementItem().arrangementId(r.getExternalArrangementId()))
+                .collectList()
+                .flatMap(this::refreshTransactions)
+                .flux();
+        }
+        return res.flatMap(this::ingestTransactions);
+    }
+
+    @Override
     public Mono<?> ingestTransactionsAsync(TransactionPullIngestionRequest res) {
         if (properties.getChains().getTransactionManager().getEnabled()) {
             return Mono.just(res)
                 .publishOn(Schedulers.boundedElastic())
-                .doOnNext(request -> refreshTransactions(request).subscribe())
+                .map(request -> new ArrangementItem().arrangementId(request.getExternalArrangementId()))
+                .doOnNext(request -> refreshTransactions(List.of(request)).subscribe())
                 .doOnNext(
                     t -> log.info("Async transaction ingestion called for arrangement: {}", res.getArrangementId()));
         }
@@ -58,10 +78,27 @@ public class TransactionIngestionServiceImpl implements TransactionIngestionServ
                 t.getEvent().getArrangementId()));
     }
 
-    private Mono<Void> refreshTransactions(TransactionPullIngestionRequest res) {
-        var arrangement = new ArrangementItem().arrangementId(res.getExternalArrangementId());
-        return transactionManagerApi.postRefresh(List.of(arrangement), null, null, null)
-            .doOnSuccess(r -> log.debug("Refreshed transactions for account {}", res.getExternalArrangementId()));
+    @Override
+    public Flux<?> ingestTransactionsAsync(Flux<TransactionPullIngestionRequest> res) {
+        if (properties.getChains().getTransactionManager().getEnabled()
+            && !properties.getChains().getTransactionManager().getSplitPerArrangement()) {
+
+            return res.map(r -> new ArrangementItem().arrangementId(r.getExternalArrangementId()))
+                .publishOn(Schedulers.boundedElastic())
+                .collectList()
+                .doOnNext(request -> refreshTransactions(request).subscribe())
+                .doOnNext(t -> log.info("Async transaction ingestion called for one or more arrangements"))
+                .flux();
+        }
+        return res.flatMap(this::ingestTransactionsAsync);
+    }
+
+    private Mono<Void> refreshTransactions(List<ArrangementItem> res) {
+        return transactionManagerApi.postRefresh(res, null, null, null)
+            .doOnSuccess(r -> log.debug("Refreshed transactions for accounts {}",
+                res.stream()
+                    .map(ArrangementItem::getArrangementId)
+                    .collect(Collectors.joining(","))));
     }
 
 }