+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - Product Composition now supports transaction-manager's out-of-the-box pulling mechanism, via the `transaction-pull-integration-service`. A separate chain was created for that, enabled via: `backbase.stream.compositions.product.chains.transaction-manager.enabled`
   - Only one of the transaction chains can be enabled, either the `transaction-manager` or `transaction-composition`.
   - The purpose of this new chain is to create better support for the OOTB implementation and enable ModelBank projects.
+  - The refresh supports fine graining the pulling requests per arrangement, or a bulk request for all arrangements at once via configuring `splitPerArrangement` and `concurrency` properties.
 - Multi-tenancy support to SSDK message broker Events
   - Only supported for `spring.cloud.stream.default.consumer.concurrency=1` - Will be enhanced when upgraded to Service SDK 16
   - **Breaking Change**: Property `backbase.stream.client.headersToForward` is now replaced by `backbase.stream.context.headersToForward`